+++ src/shell/views/Shell/Shell.js
@@ -10,7 +10,7 @@ import AppError from "shell/components/AppError";
 import GlobalSidebar from "shell/components/global-sidebar";
 import GlobalTopbar from "shell/components/GlobalTopbar";
 import Missing from "shell/components/missing";
-import { SearchPage } from "shell/views/SearchPage";
+import SearchPage from "shell/views/SearchPage";
 
 import ContentApp from "apps/content-editor/src";
 import DamApp from "apps/media/src";
